pyproject.toml
[project]
name = "sheets-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chalice",
    "google-auth>=2.48.0",
    "requests>=2.32.5",
]
